Passing first test case
Added a "referenceResolver" that will attempt to resolve all references
at the end of parsing when all objects will have been created. This is
because the order of in which properties are evaluated is not guaranteed
to be the same as when the AST was serialized.

Needs more testing and more class remapping.